add scripts to update the dird_tenants from the old dird config
before stopping the services the source config is dumped to /var/lib/xivo-upgrade/wazo_dird_source.json
the dump script reads the config file hierarchy of wazo-dird or xivo-dird which will execute xivo-confgen
to generate the sources.

After the upgrade, all dird_sources in the DB have a uuid field but the backend is still unknown. The migration
will have to update existing sources and create missing sources based on the config.

Updating the existing sources is important to maintain the favorite foreign key in dird_favorites