dird migration: remove the version check
the version prevents the script to be executed if it did not get executed on
the upgrade from < 19.08.

removing the check will allow the script to be executed if the administrator
re-starts wazo-upgrade to fix the upgrade

the script is meant to be executed multiple times, if all the resource already
exits the script will still complete successfully.